requirements.txt
langchain==0.1.0
langchain-community==0.0.10
langchain-openai==0.0.2
chainlit==0.7.700
python-dotenv==1.0.0
qdrant-client==1.14.3
pypdf==3.17.4
pymupdf==1.23.14
openai==1.6.1
tiktoken==0.5.2
numpy==1.24.3
pandas==2.1.4
aiofiles==23.2.1
pydantic==2.5.2